+++ packages/lumx-core/src/js/types/jsx/PropsToOverride.ts
@@ -3,6 +3,7 @@ export type PropsToOverride =
     | 'ref'
     | 'handleClick'
     | 'handleChange'
+    | 'handleInput'
     | 'handleKeyPress'
     | 'handleClose'
     | 'handleFocus'